rename: png 파일 이름 svg로 수정
